Udemy: dodane logowanie i uzytkownicy

+++ Udemy/YourFirstRESTAPI/resources/item.py
@@ -1,7 +1,7 @@
-import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 
 from db import db
@@ -17,14 +17,14 @@ class ItemsList(MethodView):
         return ItemModel.query.all()
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class ItemID(MethodView):
     @blp.response(200, ItemSchema)
     def get(self,item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
-
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(200, ItemSchema)
     def put(self,item_data,item_id):
@@ -39,7 +39,12 @@ class ItemID(MethodView):
         db.session.commit()
         return item
 
+    @jwt_required()
     def delete(self,item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -48,6 +53,7 @@ class ItemID(MethodView):
 
 @blp.route("/item")
 class Item(MethodView):
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self,  item_data):
